build: require Automake-NG
* configure.ac (AC_PREREQ): Require Autoconf version >= 2.65,
since that is the minimal version supported by Automake-NG.
(AM_INIT_AUTOMAKE): Add the 'ng' option, to ensure that mainstream
Automake is not used by mistake when bootstrapping.  Also, bump
the required Automake version from '1.11' to '1.12a', which is the
latest (and still development-only) version of Automake-NG at the
moment of writing.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>