build: don't remove GNUmakefile in non-VPATH builds
This fix is needed for Automake-NG.  It is basically an almost-verbatim
port of this Gnulib patch:
<http://lists.gnu.org/archive/html/bug-gnulib/2012-08/msg00014.html>

* Makefile.am (distclean-local): The makefiles generated by Automake-NG
always contain a definition of VPATH, even in non-VPATH builds (its value
being simply '.' in that case).  So, to determine whether running under a
VPATH setup, compare '$(srcdir)' to '.' rather than checking whether
'$(VPATH)' expands to the empty string.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>